* DefObject.h (setBit, getField): Use unsigned long instead of unsigned. (int is smaller than void * on alphaev56-dec-osf4.0.)
git-svn-id: svn://svn.savannah.nongnu.org/swarm/trunk@4536 84d386b0-d280-4341-8980-c4cb7e1eb7a5

+++ swarm/src/defobj/DefObject.h
@@ -93,15 +93,15 @@ extern id lispInKeyword (id index);
 //
 
 #define getBit(word, bit) ((word) & bit)
-#define setBit(word, bit, value)((value) ? ((word) |= (unsigned)bit) : \
-                                           ((word) &= ~((unsigned)bit)))
+#define setBit(word, bit, value)((value) ? ((word) |= (unsigned long)bit) : \
+                                           ((word) &= ~((unsigned long)bit)))
 
 // ((value) ? ((word) |= bit) : 
 //				  ((word) &= ~ bit))
 
 #define setField(word, shift, value) (word) |= ((value) << shift)
 
-#define getField(word, shift, mask) ((unsigned)((word) & mask) >> shift)
+#define getField(word, shift, mask) ((unsigned long)((word) & mask) >> shift)
 
 //
 // callMethodInClass() -- macro for method lookup in an alternate superclass
